Add `PartialCycle::surface`
