+++ crates/fj-kernel/src/algorithms/reverse/face.rs
@@ -2,6 +2,7 @@ use crate::{
     insert::Insert,
     objects::{Face, Objects},
     partial::HasPartial,
+    partial2::{FullToPartialCache, Partial},
     services::Service,
     storage::Handle,
 };
@@ -10,10 +11,17 @@ use super::Reverse;
 
 impl Reverse for Handle<Face> {
     fn reverse(self, objects: &mut Service<Objects>) -> Self {
-        let exterior = self.exterior().clone().reverse(objects);
+        let mut cache = FullToPartialCache::default();
+
+        let exterior = Partial::from_full(
+            self.exterior().clone().reverse(objects),
+            &mut cache,
+        );
         let interiors = self
             .interiors()
-            .map(|cycle| cycle.clone().reverse(objects))
+            .map(|cycle| {
+                Partial::from_full(cycle.clone().reverse(objects), &mut cache)
+            })
             .collect::<Vec<_>>();
 
         Face::partial()
